Split user_script from webview.h
