+++ core/include/webview/detail/backends/cocoa_webkit.hh
@@ -37,6 +37,7 @@
 //
 
 #include "../../types.hh"
+#include "../user_script.hh"
 
 #include <CoreGraphics/CoreGraphics.h>
 #include <objc/NSObjCRuntime.h>
